navigationToWatchPage test added
