+++ src/test/java/org/espn/tests/ActionsForActiveUserTest.java
@@ -1,0 +1,36 @@
+package org.espn.tests;
+
+import org.espn.pages.UserOptionsIFrame;
+import org.espn.pages.LoginIFrame;
+import org.espn.pages.WatchPage;
+import org.testng.annotations.Test;
+
+import static org.hamcrest.Matchers.is;
+
+public class ActionsForActiveUserTest extends BaseTest {
+    @Test(dataProvider = "usersLoginData-provider", priority = 1)
+    public void login(String email, String password) {
+        UserOptionsIFrame userOptionsIFrame = super.basePage.goToUserOptions();
+        LoginIFrame loginIFrame = userOptionsIFrame.actionForClickLoginButton();
+        checkThat("modal is present", loginIFrame.areLoginFormElementsDisplayed(), is(true));
+        loginIFrame.triggerUserLogging(email, password);
+        userOptionsIFrame = super.basePage.goToUserOptions();
+        checkThat("login succeeded", userOptionsIFrame.getUsernameLogged(), is("Ana!"));
+        userOptionsIFrame.goFromUserOptionsToBasePage();
+    }
+
+    @Test(priority = 2)
+    public void navigationToWatchPage() {
+        WatchPage watchPage = super.basePage.goToWatchPage();
+        checkThat("more than one carousel is present", watchPage.areWatchPageElementsDisplayed()
+                , is(true));
+        watchPage.clickOnCarouselSecondCard();
+        checkThat("choose supplier frame is present", watchPage.isExitFromChooseSupplierBtnDisplayed()
+                , is(true));
+        watchPage.clickForExitFromChooseSupplier();
+        watchPage.goToBasePage();
+        UserOptionsIFrame userOptionsIFrame = super.basePage.goToUserOptions();
+        checkThat("user still connected", userOptionsIFrame.getUsernameLogged(), is("Ana!"));
+        userOptionsIFrame.goFromUserOptionsToBasePage();
+    }
+}
